requirements.txt
paho-mqtt
requests
pep8
pylint
mock
pyfakefs
twine
pypandoc
httmock